bbcradio.py: Standalone ('if main') code removed
bbcradio.py should not (and really now can not) be
run as a standalone script. Instead it should be
imported by dedicated scripts and used that way.
To that effect, the 'if main' block at the end of
the file has been removed.

Also removed import statements for two unused
modules (os, datetime.date) - they haven’t been
needed since the function to write out a track
list was removed in b8150f0ae572